python/requirements.txt
Twisted==15.0.0
argparse==1.2.1
thrift==0.9.2
wsgiref==0.1.2
zope.interface==4.1.2
